+++ covasim/interventions.py
@@ -608,11 +608,11 @@ class test_num(Intervention):
         if self.pdf: # Handle the onset to swab delay
             symp_time = int(t - sim.people.date_symptomatic[symp_inds]) # Find time since symptom onset
             inv_count = (np.bincount(symp_time)/len(symp_time)) # Find how many people have had symptoms of a set time and invert
-            count = np.nan * np.ones(inv_count.shape) #
-            count[inv_count != 0] = 1/inv_count[inv_count != 0]
-            symp_test *= self.pdf.pdf(symp_time) * count[symp_time] # Put it all together, have to add a small amount because 0.0 will fail
+            count = np.nan * np.ones(inv_count.shape) # Initialize the count
+            count[inv_count != 0] = 1/inv_count[inv_count != 0] # Update the counts where defined
+            symp_test *= self.pdf.pdf(symp_time) * count[symp_time] # Put it all together
 
-        test_probs[symp_inds] *= symp_test
+        test_probs[symp_inds] *= symp_test # Update the test probabilities
 
         # Handle symptomatic testing, taking into account prevalence of ILI symptoms
         if self.ili_prev is not None:
@@ -666,7 +666,7 @@ class test_prob(Intervention):
         interv = cv.test_prob(symp_quar_prob=0.4) # Test 40% of those in quarantine with symptoms
     '''
     def __init__(self, symp_prob, asymp_prob=0.0, symp_quar_prob=None, asymp_quar_prob=None, subtarget=None, ili_prev=None,
-                 test_sensitivity=1.0, loss_prob=0.0, test_delay=0, start_day=0, end_day=None, **kwargs):
+                 test_sensitivity=1.0, loss_prob=0.0, test_delay=0, start_day=0, end_day=None, swab_delay=None, **kwargs):
         super().__init__(**kwargs) # Initialize the Intervention object
         self._store_args() # Store the input arguments so the intervention can be recreated
         self.symp_prob        = symp_prob
@@ -680,6 +680,7 @@ class test_prob(Intervention):
         self.test_delay       = test_delay
         self.start_day        = start_day
         self.end_day          = end_day
+        self.pdf              = cvu.get_pdf(**sc.mergedicts(swab_delay)) # If provided, get the distribution's pdf -- this returns an empty dict if None is supplied
         return
 
 
@@ -689,9 +690,7 @@ class test_prob(Intervention):
         self.end_day   = sim.day(self.end_day)
         self.days      = [self.start_day, self.end_day]
         self.ili_prev  = process_daily_data(self.ili_prev, sim, self.start_day)
-
         self.initialized = True
-
         return
 
 
@@ -703,29 +702,43 @@ class test_prob(Intervention):
         elif self.end_day is not None and t > self.end_day:
             return
 
-        # Define symptomatics, accounting for ILI prevalence
+        # Find probablity for symptomatics to be tested
         symp_inds  = cvu.true(sim.people.symptomatic)
+        symp_prob = self.symp_prob
+        if self.pdf:
+            symp_time = int(t - sim.people.date_symptomatic[symp_inds]) # Find time since symptom onset
+            inv_count = (np.bincount(symp_time)/len(symp_time)) # Find how many people have had symptoms of a set time and invert
+            count = np.NaN * np.ones(inv_count.shape)
+            count[inv_count != 0] = 1/inv_count[inv_count != 0]
+            symp_prob = np.ones(len(symp_time))
+            inds = 1 > (symp_time*self.symp_prob)
+            symp_prob[inds] = self.symp_prob/(1-symp_time[inds]*self.symp_prob)
+            symp_prob = self.pdf.pdf(symp_time) * symp_prob * count[symp_time]
+
+        # Define symptomatics, accounting for ILI prevalence
+        ili_inds = []
         if self.ili_prev is not None:
             rel_t = t - self.start_day
             if rel_t < len(self.ili_prev):
                 n_ili = int(self.ili_prev[rel_t] * sim['pop_size'])  # Number with ILI symptoms on this day
                 ili_inds = cvu.choose(sim['pop_size'], n_ili) # Give some people some symptoms, assuming that this is independent of COVID symptomaticity...
-                symp_inds = np.unique(np.concatenate((symp_inds, ili_inds)),0)
+                ili_inds = np.setdiff1d(ili_inds, symp_inds)
 
         # Define asymptomatics: those who neither have COVID symptoms nor ILI symptoms
-        asymp_inds = np.setdiff1d(np.arange(sim['pop_size']), symp_inds)
+        asymp_inds = np.setdiff1d(np.setdiff1d(np.arange(sim['pop_size']), symp_inds), ili_inds)
 
         # Handle quarantine and other testing criteria
         quar_inds       = cvu.true(sim.people.quarantined)
         symp_quar_inds  = np.intersect1d(quar_inds, symp_inds)
         asymp_quar_inds = np.intersect1d(quar_inds, asymp_inds)
+        diag_inds       = cvu.true(sim.people.diagnosed)
         if self.subtarget is not None:
             subtarget_inds  = self.subtarget['inds']
-        diag_inds       = cvu.true(sim.people.diagnosed)
 
         # Construct the testing probabilities piece by piece -- complicated, since need to do it in the right order
         test_probs = np.zeros(sim.n) # Begin by assigning equal testing probability to everyone
-        test_probs[symp_inds]       = self.symp_prob       # People with symptoms
+        test_probs[symp_inds]       = symp_prob            # People with symptoms
+        test_probs[ili_inds]        = self.symp_prob       # People with symptoms
         test_probs[asymp_inds]      = self.asymp_prob      # People without symptoms
         test_probs[symp_quar_inds]  = self.symp_quar_prob  # People with symptoms in quarantine
         test_probs[asymp_quar_inds] = self.asymp_quar_prob # People without symptoms in quarantine
@@ -735,7 +748,7 @@ class test_prob(Intervention):
         test_probs[diag_inds] = 0.0 # People who are diagnosed don't test
         test_inds = cvu.binomial_arr(test_probs).nonzero()[0] # Finally, calculate who actually tests
 
-        sim.people.test(test_inds, test_sensitivity=self.test_sensitivity, loss_prob=self.loss_prob, test_delay=self.test_delay)
+        sim.people.test(test_inds, test_sensitivity=self.test_sensitivity, loss_prob=self.loss_prob, test_delay=self.test_delay) # Actually test people
         sim.results['new_tests'][t] += int(len(test_inds)*sim['pop_scale']/sim.rescale_vec[t]) # If we're using dynamic scaling, we have to scale by pop_scale, not rescale_vec
 
         return